Command-line arguments

readRequiredFilenameARG — Function

readRequiredFilenameARG(myARGS, currfilename, callingfunc)

Read from the list of command-line arguments myARGS the file name prefixed by currfilename and an equals sign. Throw an error (using callingfunc to emphasize where it is needed) if that argument is missing.

See also: readOptionalFilenameARG, readGenvstr, readOptionalBoolcondARG, readOptionalFloatType, readfreqlist!, readklist

readOptionalFilenameARG — Function

readOptionalFilenameARG(myARGS, currfilename)

Read from the list of command-line arguments myARGS the file name prefixed by currfilename and an equals sign. Return a blank string if that argument is missing.

See also: readRequiredFilenameARG, readGenvstr, readOptionalBoolcondARG, readOptionalFloatType, readfreqlist!, readklist

readGenvstr — Function

readGenvstr(myARGS)

Read from the list of command-line arguments myARGS the environmental Green's function prefixed by "Genv=". Return "PEC" (representing a PEC plane) only if that is explicitly found, otherwise return "VAC" by default.

See also: readRequiredFilenameARG, readOptionalFilenameARG, readOptionalBoolcondARG, readOptionalFloatType, readfreqlist!, readklist

readOptionalBoolcondARG — Function

readOptionalBoolcondARG(myARGS, currvarname, outfilename)

Read from the list of command-line arguments myARGS optional boolean arguments labeled by currvarname, and modify outfilename as appropriate.

See also: readRequiredFilenameARG, readOptionalFilenameARG, readGenvstr, readOptionalFloatType, readfreqlist!, readklist

readOptionalFloatType — Function

readOptionalFloatType(myARGS)

Read from the list of command-line arguments myARGS optional arguments labeling the floating-point type. Return Float64 by default if nothing else is found.

See also: readRequiredFilenameARG, readOptionalFilenameARG, readGenvstr, readOptionalBoolcondARG, readfreqlist!, readklist

readfreqlist! — Function

readfreqlist!(freqlist, myARGS, callingfunc)

Read from the list of command-line arguments myARGS the list of frequencies, modifying freqlist "in-place" (actually appending as needed), and using the name callingfunc to be more informative when throwing errors.

Note

Any frequencies will be converted to their absolute values. Frequencies are not allowed to be exactly zero.

See also: readRequiredFilenameARG, readOptionalFilenameARG, readGenvstr, readOptionalBoolcondARG, readOptionalFloatType, readklist

readklist — Function

readklist(myARGS, callingfunc, numdims)

Read from the list of command-line arguments myARGS the list of Bloch wavevectors, returning zeros(FT, 1, 3) if numdims is zero.

See also: readRequiredFilenameARG, readOptionalFilenameARG, readGenvstr, readOptionalBoolcondARG, readOptionalFloatType, readfreqlist!