Exponential integral function

These functions can all be used independently of AARMBEM.jl

E1func — Function

E1func(x)

Compute the exponential integral function $E_{1}(x)$ for a real number x. This function returns zero if x is larger than 700, or Inf + pi*1im if x is smaller than -700. (This definition also defines the branch taken in the complex plane for $x < 0$.) Note that even though $E_{1}(x)$ is real for $x > 0$, the result of E1func(x) is cast as a complex type, based on the type of x, for consistency.

This is a reimplementation of code at https://github.com/mitmath/18S096/blob/iap2017/pset3/pset3-solutions.ipynb by Steven G. Johnson, based on a combination of Taylor series and continued fraction expansions.

See also: E1taylor, E1cf

E1taylor — Function

E1taylor(x, maxorder)

Compute the exponential integral function $E_{1}(x)$ for a real number x, using the Taylor series expansion up to a maximum order maxorder. This function works best for $-30 \leq x \leq 3$.

This is a reimplementation of code at https://github.com/mitmath/18S096/blob/iap2017/pset3/pset3-solutions.ipynb by Steven G. Johnson.

See also: E1func, E1cf

E1cf — Function

E1cf(x, maxorder)

Compute the exponential integral function $E_{1}(x)$ for a real number x, using the continued fraction expansion up to a maximum order maxorder. This function works best for $x < -30$ or $x > 3$.

This is a reimplementation of code at https://github.com/mitmath/18S096/blob/iap2017/pset3/pset3-solutions.ipynb by Steven G. Johnson.

See also: E1func, E1taylor