Generic matrix and vector utilities

These functions can all be used independently of AARMBEM.jl.

Involving a full matrix and a diagonal matrix as a vector

addtodiag! — Function

addtodiag!(A, v)

Efficiently add the vector v in-place to the main diagonal of the matrix A (overwriting the latter). length(v) must match one of the dimensions of A, with the other dimension being at least as large, and the element types of v must be promotable to those of A. If so, yield the same result as A .= diagm(v) .+ A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> addtodiag!(A, v)
3×3 Array{Float64,2}:
 2.0  2.0   3.0
 2.0  6.0   6.0
 3.0  6.0  12.0

See also: addtodiag, subtractfromdiag!, subtractfromdiag

addtodiag — Function

addtodiag(A, v)

Efficiently add the vector v not in-place to the main diagonal of the matrix A, returning the result. length(v) must match one of the dimensions of A, with the other dimension being at least as large, and the element types of v must be promotable to those of A. If so, yield the same result as diagm(v) .+ A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> addtodiag(A, v)
3×3 Array{Float64,2}:
 2.0  2.0   3.0
 2.0  6.0   6.0
 3.0  6.0  12.0

See also: addtodiag!, subtractfromdiag!, subtractfromdiag

subtractfromdiag! — Function

subtractfromdiag!(A, v)

Efficiently subtract the main diagonal of the matrix A from the vector v in-place (overwriting the latter). length(v) must match one of the dimensions of A, with the other dimension being at least as large, and the element types of v must be promotable to those of A. If so, yield the same result as A .= diagm(v) .- A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> subtractfromdiag!(A, v)
3×3 Array{Float64,2}:
  0.0  -2.0  -3.0
 -2.0  -2.0  -6.0
 -3.0  -6.0  -6.0

See also: addtodiag!, addtodiag, subtractfromdiag

subtractfromdiag — Function

subtractfromdiag(A, v)

Efficiently subtract the main diagonal of the matrix A from the vector v not in-place, returning the result. length(v) must match one of the dimensions of A, with the other dimension being at least as large, and the element types of v must be promotable to those of A. If so, yield the same result as diagm(v) .- A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> subtractfromdiag(A, v)
3×3 Array{Float64,2}:
 1.0  2.0  3.0
 2.0  4.0  6.0
 3.0  6.0  9.0

See also: addtodiag!, addtodiag, subtractfromdiag!

mulMatDiagleft! — Function

mulMatDiagleft!(A, v)

Efficiently multiply each row of the matrix A by the corresponding element of the vector v in-place (overwriting the former). The element types of v must be promotable to those of A. If so, yield the same result as A .= diagm(v) * A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> mulMatDiagleft!(A, v)
3×3 Array{Float64,2}:
 1.0   2.0   3.0
 4.0   8.0  12.0
 9.0  18.0  27.0

See also: mulMatDiagleft, mulMatDiagright!, mulMatDiagright

mulMatDiagleft — Function

mulMatDiagleft(A, v)

Efficiently multiply each row of the matrix A by the corresponding element of the vector v not in-place, returning the result. The element types of v must be promotable to those of A. If so, yield the same result as diagm(v) * A.

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> mulMatDiagleft(A, v)
3×3 Array{Float64,2}:
 1.0   2.0   3.0
 4.0   8.0  12.0
 9.0  18.0  27.0

See also: mulMatDiagleft!, mulMatDiagright!, mulMatDiagright

mulMatDiagright! — Function

mulMatDiagright!(A, v)

Efficiently multiply each column of the matrix A by the corresponding element of the vector v in-place (overwriting the former). The element types of v must be promotable to those of A, and v must have the same length as the number of columns of A. If so, yield the same result as A .= A * diagm(v).

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> mulMatDiagright!(A, v)
3×3 Array{Float64,2}:
 1.0   4.0   9.0
 2.0   8.0  18.0
 3.0  12.0  27.0

See also: mulMatDiagleft!, mulMatDiagleft, mulMatDiagright

mulMatDiagright — Function

mulMatDiagright(A, v)

Efficiently multiply each column of the matrix A by the corresponding element of the vector v not in-place, returning the result. The element types of v must be promotable to those of A. If so, yield the same result as A * diagm(v).

Examples

julia> A = [1.0 2.0 3.0; 2.0 4.0 6.0; 3.0 6.0 9.0];

julia> v = [1.0; 2.0; 3.0];

julia> mulMatDiagright(A, v)
3×3 Array{Float64,2}:
 1.0   4.0   9.0
 2.0   8.0  18.0
 3.0  12.0  27.0

See also: mulMatDiagleft!, mulMatDiagleft, mulMatDiagright!

Involving only a vector

vecNto3N — Function

vecNto3N(v)

Return a vector w that is 3 times the length of v, such that each 3-element block w[3*ii-2:3*ii] has all elements identical to v[ii]. Mathematically, the vector

\[v = \begin{bmatrix} v_1 & v_2 & \ldots & v_N \end{bmatrix}^{\top}\]

is mapped to the vector

\[w = \begin{bmatrix} v_1 & v_1 & v_1 & v_2 & v_2 & v_2 & \ldots & v_N & v_N & v_N \end{bmatrix}^{\top}\]

Examples

julia> v = [1.0; 2.0; 3.0];

julia> vecNto3N(v)
9-element Array{Float64,1}:
 1.0
 1.0
 1.0
 2.0
 2.0
 2.0
 3.0
 3.0
 3.0

Involving only a square matrix

sympart! — Function

sympart!(A)

Convert a square matrix A into its symmetric part in-place (overwriting A), yielding the same result as A .= (A .+ transpose(A))./2.

Examples

julia> A = [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> B = 1im .* A;

julia> sympart!(A)
3×3 Array{Float64,2}:
 1.0  3.0  5.0
 3.0  5.0  7.0
 5.0  7.0  9.0

julia> sympart!(B)
3×3 Array{Complex{Float64},2}:
 0.0+1.0im  0.0+3.0im  0.0+5.0im
 0.0+3.0im  0.0+5.0im  0.0+7.0im
 0.0+5.0im  0.0+7.0im  0.0+9.0im

See also: sympart, hermpart!, hermpart, ahermpart!, ahermpart

sympart — Function

sympart(A)

Convert a square matrix A into its symmetric part not in-place, returning the result, and yielding the same result as (A .+ transpose(A))./2.

Examples

julia> A = [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> B = 1im .* A;

julia> sympart(A)
3×3 Array{Float64,2}:
 1.0  3.0  5.0
 3.0  5.0  7.0
 5.0  7.0  9.0

julia> sympart(B)
3×3 Array{Complex{Float64},2}:
 0.0+1.0im  0.0+3.0im  0.0+5.0im
 0.0+3.0im  0.0+5.0im  0.0+7.0im
 0.0+5.0im  0.0+7.0im  0.0+9.0im

See also: sympart!, hermpart!, hermpart, ahermpart!, ahermpart

hermpart! — Function

hermpart!(A)

Convert a square matrix A into its Hermitian part in-place (overwriting A), yielding the same result as A .= (A .+ adjoint(A))./2. (This automatically uses sympart!() if A is real to avoid conflict with promotion to complex types.)

Examples

julia> A = [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> B = 1im .* A;

julia> hermpart!(A)
3×3 Array{Float64,2}:
 1.0  3.0  5.0
 3.0  5.0  7.0
 5.0  7.0  9.0

julia> hermpart!(B)
3×3 Array{Complex{Float64},2}:
 0.0+0.0im  0.0-1.0im  0.0-2.0im
 0.0+1.0im  0.0+0.0im  0.0-1.0im
 0.0+2.0im  0.0+1.0im  0.0+0.0im

See also: sympart!, sympart, hermpart, ahermpart!, ahermpart

hermpart — Function

hermpart(A)

Convert a square matrix A into its Hermitian part not in-place, returning the result, and yielding the same result as (A .+ adjoint(A))./2. (Automatically use sympart() if A is real.)

Examples

julia> A = [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> B = 1im .* A;

julia> hermpart(A)
3×3 Array{Float64,2}:
 1.0  3.0  5.0
 3.0  5.0  7.0
 5.0  7.0  9.0

julia> hermpart(B)
3×3 Array{Complex{Float64},2}:
 0.0+0.0im  0.0-1.0im  0.0-2.0im
 0.0+1.0im  0.0+0.0im  0.0-1.0im
 0.0+2.0im  0.0+1.0im  0.0+0.0im

See also: sympart!, sympart, hermpart!, ahermpart!, ahermpart

ahermpart! — Function

ahermpart!(A)

Convert a square matrix A into its antiHermitian part in-place (overwriting A), yielding the same result as A .= (A .- adjoint(A))./2im. (Note that the result is Hermitian.)

Examples

julia> B = 1im .* [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> ahermpart!(B)
3×3 Array{Complex{Float64},2}:
 1.0+0.0im  3.0-0.0im  5.0-0.0im
 3.0+0.0im  5.0+0.0im  7.0-0.0im
 5.0+0.0im  7.0+0.0im  9.0+0.0im

See also: sympart!, sympart, hermpart!, hermpart, ahermpart

ahermpart — Function

ahermpart(A)

Convert a square matrix A into its antiHermitian part not in-place, returning the result, and yielding the same result as (A .- adjoint(A))./2im. (Note that the result is Hermitian.)

Examples

julia> B = 1im .* [1.0 2.0 3.0; 4.0 5.0 6.0; 7.0 8.0 9.0];

julia> ahermpart(B)
3×3 Array{Complex{Float64},2}:
 1.0+0.0im  3.0-0.0im  5.0-0.0im
 3.0+0.0im  5.0+0.0im  7.0-0.0im
 5.0+0.0im  7.0+0.0im  9.0+0.0im

See also: sympart!, sympart, hermpart!, hermpart, ahermpart!