Maxwell Green's functions

c — Constant

The vacuum speed of light is a global constant (299792458 meters per second).

Vacuum Green's function

This section of functions can be used outside of AARMBEM.jl.

GFVACGG! — Function

GFVACGG!(GvacGG::AbstractArray{Complex{FT}, 2},
         startidx1::Integer, startidx2::Integer,
         freq::Union{FT, Complex{FT}}, blochk::Array{FT, 1},
         posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
         latticevecs::Array{FT, 2}, reciprocalvecs::Array{FT, 2},
         alphafield::FT, alphasource::FT,
         numdims::Integer) where FT<:AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, using the appropriate function for numdims periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). numdims must be 0, 1, or 2.

GFVACGG!(GvacGG::AbstractArray{Complex{FT}, 2}, startidx1::Integer, startidx2::Integer,
         freq::Union{FT, Complex{FT}},
         posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
         alphafield::FT, alphasource::FT) where {FT<:AbstractFloat}

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 Green's function interaction tensor in vacuum at frequency freq, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, for nonperiodic geometries.

GFVACGG — Function

GFVACGG(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
        posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
        latticevecs::AbstractArray{FT, 2}, reciprocalvecs::AbstractArray{FT, 2},
        alphafield::FT, alphasource::FT, numdims::Integer) where FT<:AbstractFloat

Return the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, using the appropriate function for numdims periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). numdims must be 0, 1, or 2.

GFVACGG(freq::Union{FT, Complex{FT}},
        posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
        alphafield::FT, alphasource::FT) where {FT<:AbstractFloat}

Return the 3-by-3 Green's function interaction tensor in vacuum at frequency freq, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, for nonperiodic geometries.

GFVACGG_sca! — Function

GFVACGG_sca!(GvacGG::AbstractArray{Complex{FT}, 2},
             startidx1::Integer, startidx2::Integer,
             freq::Union{FT, Complex{FT}}, blochk::Array{FT, 1},
             posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
             latticevecs::Array{FT, 2}, reciprocalvecs::Array{FT, 2},
             alphafield::FT, alphasource::FT,
             numdims::Integer) where FT <: AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with a 3-by-3 block of zeros, which is the scattering Green's function in vacuum by definition, irrespective of the other function arguments.

GFVACGG_sca!(GvacGG::AbstractArray{Complex{FT}, 2}, startidx1::Integer, startidx2::Integer,
             freq::Union{FT, Complex{FT}}, posfield::AbstractArray{FT, 1},
             possource::AbstractArray{FT, 1}, alphafield::FT,
             alphasource::FT) where {FT<:AbstractFloat}

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with a 3-by-3 block of zeros, which is the scattering Green's function in vacuum by definition, irrespective of the other function arguments.

GFVACGG_sca — Function

GFVACGG_sca(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
            posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
            latticevecs::AbstractArray{FT, 2}, reciprocalvecs::AbstractArray{FT, 2},
            alphafield::FT, alphasource::FT,
            numdims::Integer) where {FT<:AbstractFloat}

Return a 3-by-3 block of zeros, which is the scattering Green's function in vacuum by definition, irrespective of the other function arguments.

GFVACGG0! — Function

GFVACGG0!(GvacGG::AbstractArray{Complex{FT}, 2},
          startidx1::Integer, startidx2::Integer,
          freq::Union{FT, Complex{FT}},
          posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
          alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GvacGG[startidx1:startidx1+2,startidx2: startidx2+2] in-place with the 3-by-3 Green's function interaction tensor in vacuum at frequency freq without spatial periodicity, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively.

GFVACGG0 — Function

GFVACGG0(freq::Union{FT, Complex{FT}},
         posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
         alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 Green's function interaction tensor in vacuum at frequency freq without spatial periodicity, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively.

GFVACGG0_sca! — Function

GFVACGG0_sca!(GvacGG::AbstractArray{Complex{FT}, 2},
              startidx1::Integer, startidx2::Integer,
              freq::Union{FT, Complex{FT}},
              posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
              alphafield::FT, alphasource::FT) where FT <: AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with a 3-by-3 block of zeros, which is the scattering Green's function in vacuum by definition, irrespective of the other function arguments.

GFVACGG0_sca — Function

GFVACGG0_sca(freq::Union{FT, Complex{FT}}, posfield::AbstractArray{FT, 1},
             possource::AbstractArray{FT, 1}, alphafield::FT,
             alphasource::FT) where {FT<:AbstractFloat}

Return a 3-by-3 block of zeros, which is the scattering Green's function in vacuum by definition, irrespective of the other function arguments.

GFVACGGcoincident! — Function

GFVACGGcoincident!(GvacGG::AbstractArray{Complex{FT}, 2},
                   startidx1::Integer, startidx2::Integer,
                   freq::Union{FT, Complex{FT}}, alphafield::FT,
                   alphasource::FT) where FT<:AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 tensor block corresponding to vacuum Green's function interaction between two Gaussian basis functions, defined by respective atomic polarizabilities alphasource and alphafield, when their centers coincide; this quantity always has a finite imaginary part, and has a finite real part when alphafield and alphasource are not both zero (meaning at least one basis function has a finite width).

GFVACGGcoincident — Function

GFVACGGcoincident(freq::Union{FT, Complex{FT}},
                  alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Return the scalar (as the tensor is the scalar multiplied by the 3-by-3 identity) corresponding to vacuum Green's function interaction between two Gaussian basis functions, defined by respective atomic polarizabilities alphasource and alphafield, when their centers coincide; this quantity always has a finite imaginary part, and has a finite real part when alphafield and alphasource are not both zero (meaning at least one basis function has a finite width).

GFVACGG1Ewald! — Function

GFVACGG1Ewald!(GvacGG::AbstractArray{Complex{FT}, 2},
               startidx1::Integer, startidx2::Integer,
               freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
               posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
               latticevecs::AbstractArray{FT, 2},
               reciprocalvecs::AbstractArray{FT, 2},
               alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 periodic dimension (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 1 dimension.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFVACGG1Ewald — Function

GFVACGG1Ewald(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
              posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
              latticevecs::AbstractArray{FT, 2},
              reciprocalvecs::AbstractArray{FT, 2},
              alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 periodic dimension (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 1 dimension.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFVACGG2Ewald! — Function

GFVACGG2Ewald!(GvacGG::AbstractArray{Complex{FT}, 2},
               startidx1::Integer, startidx2::Integer,
               freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
               posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
               latticevecs::AbstractArray{FT, 2},
               reciprocalvecs::AbstractArray{FT, 2},
               alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GvacGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 2 periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 2 dimensions.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFVACGG2Ewald — Function

GFVACGG2Ewald(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
              posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
              latticevecs::AbstractArray{FT, 2},
              reciprocalvecs::AbstractArray{FT, 2}, alphafield::FT,
              alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 2 periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 2 dimensions.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFVACGGEwaldSR — Function

GFVACGGEwaldSR(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
               posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
               latticevecs::AbstractArray{FT, 2},
               alphafield::FT, alphasource::FT, kappa::FT,
               N1max::Integer, N2max::Integer=0) where FT<:AbstractFloat

Return the short-range Ewald summation contribution to the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 or 2 periodic dimensions (with lattice parameters latticevecs for the real lattice vectors as appropriate). The sums over the real lattice run over -1*N1max:N1max and -1*N2max:N2max along each lattice vector, and kappa is the Ewald cutoff parameter.

GFVACGGEwaldLR1D — Function

GFVACGGEwaldLR1D(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
                 posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
                 reciprocalvec::AbstractArray{FT, 1},
                 alphafield::FT, alphasource::FT, kappa::FT,
                 Mmax::Integer, Smax::Integer) where FT<:AbstractFloat

Return the long-range Ewald summation contribution to the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 periodic dimension (with lattice parameters reciprocalvec for the reciprocal lattice vector as appropriate). The sum over the reciprocal lattice runs over -1*Mmax:Mmax, the sum computing the series representation of the reciprocal lattice expression runs over 0:Smax, and kappa is the Ewald cutoff parameter.

GFVACGGEwaldLR2D — Function

GFVACGGEwaldLR2D(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
                 posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
                 reciprocalvecs::AbstractArray{FT, 2},
                 alphafield::FT, alphasource::FT, kappa::FT,
                 M1max::Integer, M2max::Integer) where FT<:AbstractFloat

Return the long-range Ewald summation contribution to the 3-by-3 Green's function interaction tensor in vacuum at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 2 periodic dimensions (with lattice parameters reciprocalvecs for the reciprocal lattice vectors as appropriate). The sum over the reciprocal lattice runs over -1*M1max:M1max and -1*M2max:M2max along each reciprocal lattice vector, and kappa is the Ewald cutoff parameter.

Scattering Green's function above a PEC plane at $z = 0$

This section of functions can be used outside of AARMBEM.jl.

GFPECGG! — Function

GFPECGG!(GscaGG::AbstractArray{Complex{FT}, 2},
         startidx1::Integer, startidx2::Integer,
         freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
         posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
         latticevecs::AbstractArray{FT, 2}, reciprocalvecs::AbstractArray{FT, 2},
         alphafield::FT, alphasource::FT, numdims::Integer) where FT<:AbstractFloat

Fill GscaGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, using the appropriate function for numdims periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). numdims must be 0, 1, or 2.

GFPECGG!(GscaGG::AbstractArray{Complex{FT}, 2},
         startidx1::Integer, startidx2::Integer,
         freq::Union{FT, Complex{FT}}, posfield::AbstractArray{FT, 1},
         possource::AbstractArray{FT, 1}, alphafield::FT,
         alphasource::FT) where {FT<:AbstractFloat}

Fill GscaGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, for nonperiodic geometries.

GFPECGG — Function

GFPECGG(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
             posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
             latticevecs::AbstractArray{FT, 2}, reciprocalvecs::AbstractArray{FT, 2},
             alphafield::FT, alphasource::FT, numdims::Integer) where FT<:AbstractFloat

Return the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, using the appropriate function for numdims periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). numdims must be 0, 1, or 2.

GFPECGG(freq::Union{FT, Complex{FT}}, posfield::AbstractArray{FT, 1},
        possource::AbstractArray{FT, 1}, alphafield::FT,
        alphasource::FT) where {FT<:AbstractFloat}

Return the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, for nonperiodic geometries.

GFPECGG0! — Function

GFPECGG0!(GscaGG::AbstractArray{Complex{FT}, 2},
          startidx1::Integer, startidx2::Integer,
          freq::Union{FT, Complex{FT}},
          posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
          alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GscaGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq without spatial periodicity, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively.

GFPECGG0 — Function

GFPECGG0(freq::Union{FT, Complex{FT}}, posfield::AbstractArray{FT, 1},
         possource::AbstractArray{FT, 1}, alphafield::FT,
         alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq without spatial periodicity, corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively.

GFPECGG1Ewald! — Function

GFPECGG1Ewald!(GscaGG::AbstractArray{Complex{FT}, 2},
               startidx1::Integer, startidx2::Integer,
               freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
               posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
               latticevecs::AbstractArray{FT, 2},
               reciprocalvecs::AbstractArray{FT, 2},
               alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GscaGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 periodic dimension (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 1 dimension.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFPECGG1Ewald — Function

GFPECGG1Ewald(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
              posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
              latticevecs::AbstractArray{FT, 2},
              reciprocalvecs::AbstractArray{FT, 2}, alphafield::FT,
              alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 1 periodic dimension (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 1 dimension.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFPECGG2Ewald! — Function

GFPECGG2Ewald!(GscaGG::AbstractArray{Complex{FT}, 2},
               startidx1::Integer, startidx2::Integer,
               freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
               posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
               latticevecs::AbstractArray{FT, 2},
               reciprocalvecs::AbstractArray{FT, 2},
               alphafield::FT, alphasource::FT) where FT<:AbstractFloat

Fill GscaGG[startidx1:startidx1+2, startidx2:startidx2+2] in-place with the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 2 periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 2 dimensions.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

GFPECGG2Ewald — Function

GFPECGG2Ewald(freq::Union{FT, Complex{FT}}, blochk::AbstractArray{FT, 1},
              posfield::AbstractArray{FT, 1}, possource::AbstractArray{FT, 1},
              latticevecs::AbstractArray{FT, 2},
              reciprocalvecs::AbstractArray{FT, 2}, alphafield::FT,
              alphasource::FT) where FT<:AbstractFloat

Return the 3-by-3 scattering Green's function interaction tensor above a perfect electrically conducting (PEC) plane (coplanar with the xy-plane) from the vacuum Green's function via image theory, at frequency freq and wavevector blochk (a 3-element vector), corresponding to source position possource and field position posfield (both 3-element vectors) described by Gaussian basis functions parameterized by real atomic polarizabilities alphasource and alphafield respectively, in 2 periodic dimensions (with lattice parameters latticevecs for the real lattice vectors and reciprocalvecs for the reciprocal lattice vectors as appropriate). This is done via Ewald summation over Gaussian basis functions in 2 dimensions.

Warning

The Ewald summation convergence parameters are hard-coded into this function, though they should work for most common cases.

Green's function matrix assembly routines

assemble1MolGvacinf! — Function

assemble1MolGvacinf!(mySingleMolData::OneMol{FT}, myPeriodicData::PeriodicData{FT},
                     freq::Union{FT, Complex{FT}},
                     blochk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the Green's function interaction matrix of a single molecule in vacuum, mySingleMolData.GvacGGinf, in-place, at frequency freq and wavevector blochk using additional information from myPeriodicData to speed up the computation. In particular, if the periodic dimensionality is 0, then this matrix is symmetric (whether freq is imaginary or not). Even if the periodic dimensionality is not 0, if freq is strictly imaginary, then this matrix is Hermitian. Both of those cases allow for computing only the lower triangle of 3-by-3 blocks. Otherwise, the full matrix must be assembled.

Warning

This function depends on mySingleMolData.alpha0 having been initialized to the desired atomic susceptibilities at the same freq and blochk, but does not check this.

assemble1MolGvacinfdiag! — Function

assemble1MolGvacinfdiag!(mySingleMolData::OneMol{FT},
                         freq::Union{FT, Complex{FT}}) where FT<:AbstractFloat

Assemble the scalar coincident Green's function interaction matrix elements of a single molecule in vacuum, mySingleMolData.GvacGGinfdiag, in-place, at frequency freq.

Warning

This function depends on mySingleMolData.alpha0 having been initialized to the desired atomic susceptibilities at the same freq (and wavevector), but does not check this.

subtract1MolalphainvGvac! — Function

subtract1MolalphainvGvac!(myMolData::OneMol{<:AbstractFloat})

Subtract the Green's function interaction matrix of a single molecule in vacuum, myMolData.GvacGGinf, in-place, from its inverse susceptibility myMolData.alphainv. Separately dispatch cases with versus without phonons.

Warning

This function depends on myMolData.alphainv having been initialized to the desired molecular susceptibility at the same freq and blochk, but does not check this.

assembleGscaMolBlock! — Function

assembleGscaMolBlock!(A::SharedArray{Complex{FT}, 2},
                      myAllMolData::MolSystem{<:OneMol{FT}},
                      myTransData::TransData{FT}, nn1::Integer, nn2::Integer,
                      startidx1::Integer, startidx2::Integer,
                      myPeriodicData::PeriodicData{FT},
                      GFSCAGG!::Function, freq::Union{FT, Complex{FT}},
                      blochk::AbstractArray{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the scattering Green's function interaction matrix (evaluated via the function GFSCAGG!) between molecules nn1 and nn2 (sources evaluated in the latter, fields evaluated in the former), at frequency freq and wavevector blochk, using atomic polarizability and coordinate data respectively from myAllMolData and myTransData, and add the result in-place to the matrix block A[startidx1:endidx1, startidx2:endidx2] where endidx1 - startidx1 + 1 is 3 times the number of atoms in molecule nn1, and likewise endidx2 - startidx2 + 1 is 3 times the number of atoms in molecule nn2. If nn1 and nn2 are the same, the matrix block is symmetric if the periodic dimensionality is 0, or Hermitian if the periodic dimensionality is nonzero but freq is purely imaginary, so only the lower triangle of blocks is assembled. Otherwise, the full set of matrix blocks is assembled.

Note

If GFSCAGG! is the scattering Green's function of vacuum (returning a 3-by-3 block of zeros), namely GFVACGG_sca! or GFVACGG0_sca!, this function does nothing, saving on computations.

Warning

Currently, only GFPECGG! and GFPECGG0! are supported for GFSCAGG!, but this is not explicitly checked, and may change in future versions of this code.

Warning

This function depends on mySingleMolData.alpha0 having been initialized to the desired atomic susceptibilities at the same freq and blochk, as well as myTransData.allAtomPos having been modified to reflect the desired transformation, but does not check these.

assembleGvacMolBlock! — Function

assembleGvacMolBlock!(A::AbstractArray{Complex{FT}, 2},
                      myAllMolData::MolSystem{<:OneMol{FT}},
                      myTransData::TransData{FT}, nn1::Integer, nn2::Integer,
                      startidx1::Integer, startidx2::Integer,
                      myPeriodicData::PeriodicData{FT},
                      freq::Union{FT, Complex{FT}},
                      blochk::AbstractArray{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the vacuum Green's function interaction matrix (evaluated via the function GFVACGG!) between molecules nn1 and nn2 (sources evaluated in the latter, fields evaluated in the former), at frequency freq and wavevector blochk, using atomic polarizability and coordinate data respectively from myAllMolData and myTransData, and add the result in-place to the matrix block A[startidx1:endidx1, startidx2:endidx2] where endidx1 - startidx1 + 1 is 3 times the number of atoms in molecule nn1, and likewise endidx2 - startidx2 + 1 is 3 times the number of atoms in molecule nn2. If nn1 and nn2 are the same, the matrix block is symmetric if the periodic dimensionality is 0, or Hermitian if the periodic dimensionality is nonzero but freq is purely imaginary, so only the lower triangle of blocks is assembled. Otherwise, the full set of matrix blocks is assembled.

Tip

Use assemble1MolGvacinf! when nn1 and nn2 are the same, then stamp that into A as needed. Only use this when nn1 and nn2 are different.

Warning

This function depends on mySingleMolData.alpha0 having been initialized to the desired atomic susceptibilities at the same freq and blochk, as well as myTransData.allAtomPos having been modified to reflect the desired transformation, but does not check these.