Molecular data structures and methods

Data structures

OneMol — Type

OneMol{FT<:AbstractFloat}

Represent properties of a single molecule, with or without phonons, as an abstract type.

See also: OneMol_WithPhonons{FT}, OneMol_NoPhonons{FT}

OneMol_WithPhonons — Type

OneMol_WithPhonons{FT} <: OneMol{FT}

Represent the properties of a single molecule of $N$ atoms, with phonons.

numatoms is an integer equal to the number of atoms $N$.

Qe, Me, Be, Ke, MI, and BI are all numatoms-element real vectors.

KI is a 3*numatoms-by-3*numatoms real matrix.

molpos0 is a 3-by-numatoms matrix where molpos0[:, p] represents the coordinates of the atom labeled $p$.

molCOM is a 3-element real vector containing the unweighted molecular center of mass.

alphainv and GvacGGinf are 3*numatoms-by-3*numatoms complex matrices respectively representing the inverse susceptibility and Green's function interaction matrix of the molecule in isolation in vacuum (allocated but uninitialized until a frequency and wavevector are specified).

alpha0 is a numatoms-element real vector of the atomic polarizabilities (allocated but uninitialized until a frequency and wavevector are specified).

alphae is an numatoms-element complex vector of the purely electronic contributions to the polarizabilities, useful mainly for vdW computations (allocated but uninitialized until a frequency and wavevector are specified).

numblocks1 & numblocks2 are the integer numbers of neighboring unit cells that have nontrivial internuclear couplings to the central unit cell (each is 1 molecules that are not periodic in that direction or have no internuclear couplings/covalent bonds with neighboring unit cells in that direction).

blocklist1 & blocklist2 are range vectors indexing each unit cell block (each has the single element 0 for molecules that are not periodic in that direction or have no internuclear couplings/covalent bonds with neighboring unit cells in that direction).

See also: OneMol{FT}, OneMol_NoPhonons{FT}

OneMol_NoPhonons — Type

OneMol_NoPhonons{FT} <: OneMol{FT}

Represent the properties of a single molecule of $N$ atoms, without phonons.

numatoms is an integer equal to the number of atoms $N$.

Qe, Me, Be, and Ke are all numatoms-element real vectors.

molpos0 is a 3-by-numatoms matrix where molpos0[:, p] represents the coordinates of atom $p$.

molCOM is a 3-element real vector containing the unweighted molecular center of mass.

alphainv is a numatoms-element complex vector representing the local inverse susceptibility of each atom (allocated but uninitialized until a frequency and wavevector are specified).

GvacGGinf is a 3*numatoms-by-3*numatoms complex matrix representing the Green's function interaction matrix of the molecule in isolation in vacuum (allocated but uninitialized until a frequency and wavevector are specified).

alphae is 1 ./ alphainv, useful mainly for vdW computations (allocated but uninitialized until a frequency and wavevector are specified).

alpha0 is abs.(alphae), representing real atomic polarizabilities (allocated but uninitialized until a frequency and wavevector are specified).

See also: OneMol{FT}, OneMol_WithPhonons{FT}

MolSystem — Type

MolSystem{T<:OneMol{<:AbstractFloat}}

Represent the properties of a collection of molecules.

uniqueMolDataArray contains data for each unique molecule.

nummols is the total number of molecules (unique or not).

numatomslist is the list of atoms in every molecule (unique or not).

cumnumatomslist is the cumulative sum of the number of atoms in every molecule, to keep track of overall indexing.

dupemolinds is the list of indices tracking duplicate molecules: -1 indicates it is the first instance of that molecule, while positive index nn (in the range 1:nummols) points to nn as the molecule to be duplicated.

uniquemolinds is complementary to dupemolinds, listing the index of each unique molecule in the set of unique molecules, or -1 if it is a duplicate molecule.

See also: OneMol{FT}, OneMol_WithPhonons{FT}, OneMol_NoPhonons{FT}

Frequency-independent data structure initialization functions

read1MolDataWithPhonons — Function

read1MolDataWithPhonons(molstrarray::Array{<:AbstractString, 1},
                        mollistfilename::AbstractString, T::DataType=Float64)

Read a section of a configuration file molstrarray (an array of strings) corresponding to molecular data and return the data corresponding to that molecule, with phonons, of type OneMol_WithPhonons{T}. Create data types to use T (satisfying T<:AbstractFloat) for storing floating-point numbers, and output any errors with reference to the configuration file mollistfilename (a string).

See also: OneMol{FT}, OneMol_WithPhonons{FT}

read1MolDataNoPhonons — Function

read1MolDataNoPhonons(molstrarray::Array{<:AbstractString, 1},
                      mollistfilename::AbstractString, T::DataType=Float64)

Read a section of a configuration file molstrarray (an array of strings) corresponding to molecular data and return the data corresponding to that molecule, without phonons, of type OneMol_NoPhonons{T}. Create data types to use T (satisfying T<:AbstractFloat) for storing floating-point numbers, and output any errors with reference to the configuration file mollistfilename (a string).

See also: OneMol{FT}, OneMol_NoPhonons{FT}

readMolSystem — Function

readMolSystem(mollistfilename::AbstractString,
              nophonons:Bool, T::DataType=Float64)

Read a configuration file mollistfilename (a string) to extract molecular data of type either OneMol_NoPhonons{T}, if nophonons is true, or OneMol_WithPhonons{T} otherwise, returning information about the composite system of type MolSystem{molType{T}}. Ensure data types use T (satisfying T<:AbstractFloat) for storing floating-point numbers.

See also: OneMol{T}, OneMol_WithPhonons{T}, OneMol_NoPhonons{T}, MolSystem{molType{T}}

Frequency-dependent polarizability calculation functions

constructMolAlpha! — Method

constructMolAlpha!(myMolData::OneMol_WithPhonons{FT},
                   myPeriodicData::PeriodicData{FT},
                   freq::Union{FT, Complex{FT}},
                   k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the inverse molecular susceptibility myMolData.alphainv in-place at frequency freq and wavevector k, the latter of which must be a 3-element real vector. With phonons, myMolData.alphainv will be a matrix, the vector of atomic polarizabilities myMolData.alpha0 will depend on the elements of myMolData.alphainv, and myMolData.alphae will be independent of these two. Without phonons, all three will be simply related to each other.

constructMolAlpha! — Method

constructMolAlpha!(myMolData::OneMol_NoPhonons{FT},
                   myPeriodicData::PeriodicData{FT},
                   freq::Union{FT, Complex{FT}},
                   k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the inverse molecular susceptibility myMolData.alphainv in-place at frequency freq and wavevector k, the latter of which must be a 3-element real vector. Without phonons, myMolData.alphainv will be a vector, and the vectors myMolData.alpha0 and myMolData.alphae will be simply related to myMolData.alphainv.

KIk — Function

KIk(myMolData::OneMol_WithPhonons{FT}, myPeriodicData::PeriodicData{FT},
    k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the k-dependence of the internuclear spring constant matrix (only if phonons are present), and return the result. Return myMolData.KI if the geometry has no periodic dimensions.

See also: constructMolAlpha!(myMolData::OneMol_WithPhonons{FT}, myPeriodicData::PeriodicData{FT}, freq::Union{FT, Complex{FT}}, k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat.