Bloch periodicity

These data structures and functions may be useful outside of AARMBEM.jl

Data structures

PeriodicData — Type

PeriodicData{FT<:AbstractFloat}

Store information relevant to spatial periodicity in a system. numdims should be 0, 1, or 2, representing the number of periodic dimensions. latticevecs and reciprocalvecs should each by 3-by-2 matrices, with each column representing a 3-element lattice or corresponding reciprocal vector, with the vector being zeros along a dimension with no periodicity.

Functions

readPeriodicData — Function

readPeriodicData(FT::DataType=Float64, periodicfilename::AbstractString="")

Construct PeriodicData from an appropriate floating-point data type FT and a file of name periodicfilename containing information about spatial periodicity; if no file is specified, produce a default set of periodic data (no spatial periodicity at all).

constructlatticefromreciprocal! — Function

constructlatticefromreciprocal!(myPeriodicData::PeriodicData{<:AbstractFloat})

Reset the lattice vectors of myPeriodicData from its reciprocal vectors.

constructreciprocalfromlattice! — Function

constructreciprocalfromlattice!(myPeriodicData::PeriodicData{<:AbstractFloat})

Reset the reciprocal vectors of myPeriodicData from its lattice vectors.