vdW interactions and thermal radiation functions

Thermal radiation

heattransfer — Function

heattransfer(myARGS)

Extract command-line arguments for computing the radiative heat transfer among a collection of molecules, then compute the radiative heat transfer appropriately.

Note

Any frequencies will be converted to their absolute values. Frequencies are not allowed to be exactly zero.

Note

This computes the integrand $\operatorname{Tr}(\operatorname{asym}(\mathbb{V}_{m}^{-1}) \mathbb{P}_{m} \mathbb{T}^{\dagger} \operatorname{asym}(\mathbb{P}_{n} \mathbb{G}^{\mathrm{env}}) \mathbb{T}\mathbb{P}_{m})$ at specified real frequencies, without doing any integration. This includes neither numerical prefactors outside of the trace, nor the Planck function.

Tip

This function can be embarrassingly parallelized over frequency and wavevector. However, the integrand at each frequency and wavevector is calculated for every transformation, as this maximizes the efficiency of the program by virtue of reusing quantities independent of geometric transformations.

heattransfer_general — Function

heattransfer_general(outfilename, myAllMolData, myTransData, myPeriodicData, Genvstr,
                     nonretarded, freqklist)

Compute the actual radiative heat transfer integrand and print it to outfilename.

assemblePhimn! — Function

assemblePhimn!(Phimn::Array{FT, 2},
               myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
               myPeriodicData::PeriodicData{FT},
               totalTmat::SharedArray{Complex{FT}, 2},
               totalGenvmat::SharedArray{Complex{FT}, 2})

Assemble the matrix Phimn, whose elements Phimn[nn1, nn2] represent the energy transfer from molecule nn1 to molecule nn2, computed from trace expressions involving the matrices totalTmat and totalGenvmat, in-place. If there is no spatial periodicity, Phimn should be symmetric. If there is spatial periodicity, the transpose of Phimn at wavevector $\vec{k}$ is the equivalent of computing Phimn at $-\vec{k}$. This assumes the presence of phonons.

assemblePhimn!(Phimn::Array{FT, 2},
               myAllMolData::MolSystem{OneMol_NoPhonons{FT}},
               myPeriodicData::PeriodicData{FT},
               totalTmat::SharedArray{Complex{FT}, 2},
               totalGenvmat::SharedArray{Complex{FT}, 2})

Assemble the matrix Phimn, whose elements Phimn[nn1, nn2] represent the energy transfer from molecule nn1 to molecule nn2, computed from trace expressions involving the matrices totalTmat and totalGenvmat, in-place. If there is no spatial periodicity, Phimn should be symmetric. If there is spatial periodicity, the transpose of Phimn at wavevector $\vec{k}$ is the equivalent of computing Phimn at $-\vec{k}$. This assumes the absence of phonons.

assembleTinvGenvDiagBlock! — Function

assembleTinvGenvDiagBlock!(totalTinvmat::SharedArray{Complex{FT}, 2},
                           totalGenvmat::SharedArray{Complex{FT}, 2},
                           myAllMolData::MolSystem{OneMol_WithPhonons{FT}},
                           myTransData::TransData{FT}, nn2::Integer,
                           currmolind2::Integer, startidx2::Integer,
                           tt::Integer, myPeriodicData::PeriodicData{FT},
                           GFSCAGG!::Function,
                           currfreqG::Union{FT, Complex{FT}},
                           currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular diagonal block of the scattering Green's function interaction matrix of molecule nn2 (whose unique molecular index currmolind2 is precomputed and passed as an argument) at frequency currfreqG and wavevector currk using GFSCAGG!, then add the vacuum contribution of molecule nn2, performing any required matrix rotations corresponding to transformation tt for molecule nn2. Stamp that added to coincident contributions into totalGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])], or that (without coincident contributions) subtracted from the inverse molecular susceptibility into totalTinvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Finally, undo any rotation corresponding to transformation tt in the inverse T-operator of molecule nn2 for reuse with future transformations. This assumes the presence of phonons.

assembleTinvGenvDiagBlock!(totalTinvmat::SharedArray{Complex{FT}, 2},
                           totalGenvmat::SharedArray{Complex{FT}, 2},
                           myAllMolData::MolSystem{OneMol_NoPhonons{FT}},
                           myTransData::TransData{FT}, nn2::Integer,
                           currmolind2::Integer,
                           startidx2::Integer,
                           tt::Integer, myPeriodicData::PeriodicData{FT},
                           GFSCAGG!::Function,
                           currfreqG::Union{FT, Complex{FT}},
                           currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular diagonal block of the scattering Green's function interaction matrix of molecule nn2 (whose unique molecular index currmolind2 is precomputed and passed as an argument) at frequency currfreqG and wavevector currk using GFSCAGG!, then add the vacuum contribution of molecule nn2, performing any required matrix rotations corresponding to transformation tt for molecule nn2. Stamp that added to coincident contributions into totalGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])], or that (without coincident contributions) subtracted from the inverse molecular susceptibility into totalTinvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Finally, undo any rotation corresponding to transformation tt in the inverse T-operator of molecule nn2 for reuse with future transformations. This assumes the absence of phonons.

assembleTinvGenvOffDiagBlock! — Function

assembleTinvGenvOffDiagBlock!(totalTinvmat::SharedArray{Complex{FT}, 2},
                              totalGenvTinvmat::SharedArray{Complex{FT}, 2},
                              myAllMolData::MolSystem{<:OneMol{FT}},
                              myTransData::TransData{FT},
                              nn1::Integer, nn2::Integer,
                              startidx1::Integer,
                              startidx2::Integer,
                              myPeriodicData::PeriodicData{FT},
                              GFSCAGG!::Function,
                              currfreqG::Union{FT, Complex{FT}},
                              currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular off-diagonal block of the scattering Green's function interaction matrix of molecules nn2 as the source and nn1 as the field (whose respective unique molecular indices currmolind1 and currmolind2 are precomputed and passed as arguments) at frequency currfreqG and wavevector currk using GFSCAGG!, multiply on the left by the electronic polarizabilities, add the vacuum contribution, and stamp that into totalGenvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])] and its negative into totalTinvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. If there is no spatial periodicity, do the same for the (unconjugated) transpose, swapping nn2 and nn1 without performing further Green's function computations.

vdW interactions

vdWenergy — Function

vdWenergy(myARGS)

Extract command-line arguments for computing the vdW interaction energy among a collection of molecules, then compute the vdW energy integrand in imaginary frequency appropriately.

Note

Any frequencies will be converted to their absolute values times 1im. Frequencies are not allowed to be exactly zero.

Note

This computes the integrand $\ln(\det(\mathbb{T}_{\infty} \mathbb{T}^{-1}))$ at specified imaginary frequencies, without doing any integration or Matsubara summation. This is typically a negative dimensionless quantity, and no factors of $2\pi$ or $\hbar$ are included.

Tip

This function can be embarrassingly parallelized over frequency and wavevector. However, the integrand at each frequency and wavevector is calculated for every transformation, as this maximizes the efficiency of the program by virtue of reusing quantities independent of geometric transformations.

Tip

The recommended way to do frequency integration or Matsubara summation is to supply a list of frequencies (typically not too many), and then interpolate the integrand to do integration or Matsubara summation after the fact.

vdWenergy_general — Function

vdWenergy_general(outfilename, myAllMolData, myTransData, myPeriodicData, Genvstr,
                  nonretarded, freqklist)

Compute the actual vdW energy integrand and print it to outfilename.

constructMolalphaealphainv! — Function

constructMolalphaealphainv!(myMolData::OneMol{FT},
                            myPeriodicData::PeriodicData{FT},
                            freq::Union{FT, Complex{FT}},
                            k::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Construct the product of the electronic susceptibility with the full inverse molecular susceptibility myMolData.alphainv in-place at frequency freq and wavevector k, the latter of which must be a 3-element real vector. With phonons, myMolData.alphainv will be a matrix, the vector of atomic polarizabilities myMolData.alpha0 will depend on the elements of myMolData.alphainv, and myMolData.alphae will be independent of these two. Without phonons, all three will be simply related to each other. This is more efficient than calling constructMolAlpha! and then performing a broadcast multiplication involving myMolData.alphae multiplying the rows of myMolData.alphainv.

constructMolalphaeGvacinf! — Function

constructMolalphaeGvacinf!(myMolData::OneMol{FT},
                           myPeriodicData::PeriodicData{FT},
                           currfreqG::Union{FT, Complex{FT}},
                           currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the Green's function interaction matrix of a molecule in vacuum myMolData.GvacGGinf, then multiplies on the left by its electronic polarizabilities in-place, at frequency currfreqG and wavevector currk passing myMolData and myPeriodicData to the subsidiary functions.

mulalphaeGBlock! — Function

mulalphaeGBlock!(A::AbstractArray{Complex{FT}, 2},
                 myAllMolData::MolSystem{<:OneMol{FT}},
                 nn1::Integer, startidx1::Integer,
                 startidx2::Integer,
                 endidx2::Integer) where FT<:AbstractFloat

Assuming A represents a Green's function interaction matrix, multiply its block A[startidx1-1+3*myAllMolData.numatomslist[nn1], startidx2:endidx2] on the left by the electronic polarizabilities of molecule nn1 from myAllMolData.

integrandvac — Function

integrandvac(myAllMolData::MolSystem{<:OneMol{FT}},
             integrandvaclist::Array{FT, 1}) where FT<:AbstractFloat

Populate the list of vacuum contributions to the vdW integrand integrandvaclist with contributions from duplicate molecules where they exist, then return the sum yielding the total contribution to the integrand in vacuum.

assemblealphaeTinvDiagBlock! — Function

assemblealphaeTinvDiagBlock!(totalalphaeTinvmat::SharedArray{Complex{FT}, 2},
                             myAllMolData::MolSystem{<:OneMol{FT}},
                             myTransData::TransData{FT}, nn2::Integer,
                             currmolind2::Integer, startidx2::Integer,
                             tt::Integer, myPeriodicData::PeriodicData{FT},
                             GFSCAGG!::Function,
                             currfreqG::Union{FT, Complex{FT}},
                             currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular diagonal block of the scattering Green's function interaction matrix of molecule nn2 (whose unique molecular index currmolind2 is precomputed and passed as an argument) at frequency currfreqG and wavevector currk using GFSCAGG!, multiply on the left by the electronic polarizabilities, then add the vacuum contribution which has already been computed as the electronic polarizabilities multiplied by the inverse T-operator of molecule nn2, performing any required matrix rotations corresponding to transformation tt for molecule nn2, and stamp that into totalalphaeTinvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Finally, undo any rotation corresponding to transformation tt in the inverse T-operator of molecule nn2 for reuse with future transformations.

assemblealphaeTinvOffDiagBlock! — Function

assemblealphaeTinvOffDiagBlock!(totalalphaeTinvmat::SharedArray{Complex{FT}, 2},
                                myAllMolData::MolSystem{<:OneMol{FT}},
                                myTransData::TransData{FT},
                                nn1::Integer, nn2::Integer,
                                startidx1::Integer,
                                startidx2::Integer,
                                myPeriodicData::PeriodicData{FT},
                                GFSCAGG!::Function,
                                currfreqG::Union{FT, Complex{FT}},
                                currk::Array{FT, 1}=zeros(FT, 3)) where FT<:AbstractFloat

Assemble the molecular off-diagonal block of the scattering Green's function interaction matrix of molecules nn2 as the source and nn1 as the field (whose respective unique molecular indices currmolind1 and currmolind2 are precomputed and passed as arguments) at frequency currfreqG and wavevector currk using GFSCAGG!, multiply on the left by the electronic polarizabilities, add the vacuum contribution, and stamp the negative of that into totalalphaeTinvmat[(startidx2-1).+(1:3*myAllMolData.numatomslist[nn2]), (startidx2-1).+(1:3*myAllMolData.numatomslist[nn2])]. Do the same for the adjoint, swapping nn2 and nn1 without performing further Green's function computations.