Transformation data structures and methods

Generic functions

The functions in this section may be useful outside of AARMBEM.jl, and can be used independently in any case.

findCOM — Function

findCOM(pos0, startatom=1, endatom=size(pos0, 2))

Return the unweighted center of mass of a block pos0[:, startatom:endatom] of coordinates, where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$.

translateCOM! — Function

translateCOM!(pos0, dispCOM, startatom=1, endatom=size(pos0, 2))

Translate the block pos0[:, startatom:endatom] of coordinates in-place (overwriting it), where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$ and dispCOM is a 3-element vector representing the displacement, such that for each p in the range, pos0[:, p] .+= dispCOM.

See also: translateCOM

translateCOM — Function

translateCOM(pos0, dispCOM, startatom=1, endatom=size(pos0, 2))

Translate the block pos0[:, startatom:endatom] of coordinates not in-place, returning the result, where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$ and dispCOM is a 3-element vector representing the displacement, such that for each p in the range, pos0[:, p] .+= dispCOM.

See also: translateCOM!

rotatecoords! — Function

rotatecoords!(pos0, rotmat, startatom=1, endatom=size(pos0, 2))

Rotate the block pos0[:, startatom:endatom] of coordinates in-place (overwriting it), where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$ and rotmat is a 3-by-3 matrix representing the (alibi) rotation, such that for each p in the range, pos0[:, p] .= rotmat * pos0[:, p].

See also: rotatecoords

Warning

This function will not check whether rotmat is real and orthogonal. It is the responsibility of the caller to ensure that rotmat represents a physical rotation; otherwise, unexpected behavior could emerge.

rotatecoords — Function

rotatecoords(pos0, rotmat, startatom=1, endatom=size(pos0, 2))

Rotate the block pos0[:, startatom:endatom] of coordinates not in-place, returning the result, where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$ and rotmat is a 3-by-3 matrix representing the rotation, such that for each p in the range, pos0[:, p] .= rotmat * pos0[:, p].

Warning

This function will not check whether rotmat is real and orthogonal. It is the responsibility of the caller to ensure that rotmat represents a physical rotation; otherwise, unexpected behavior could emerge.

See also: rotatecoords!

rotate3Nmat! — Function

rotate3Nmat!(A, rotmat, startatom=1, endatom=div(size(A, 2), 3))

Rotate the contents in the diagonal block of the complex matrix A[3*startatom-2:3*endatom, 3*startatom-2:3*endatom] by rotating each 3-by-3 block therein by the 3-by-3 real orthogonal (alibi) rotation matrix rotmat; this is done in-place, overwriting A.

See also: rotate3Nmat

Warning

This function will not check whether rotmat is real and orthogonal. It is the responsibility of the caller to ensure that rotmat represents a physical rotation; otherwise, unexpected behavior could emerge.

rotate3Nmat — Function

rotate3Nmat(A, rotmat, startatom=1, endatom=div(size(A, 2), 3))

Rotate the contents in the diagonal block of the complex matrix A[3*startatom-2:3*endatom, 3*startatom-2:3*endatom] by rotating each 3-by-3 block therein by the 3-by-3 real orthogonal (alibi) rotation matrix rotmat; this is done not in-place, returning the result.

See also: rotate3Nmat!

Warning

This function will not check whether rotmat is real and orthogonal. It is the responsibility of the caller to ensure that rotmat represents a physical rotation; otherwise, unexpected behavior could emerge.

makeCOM0! — Function

makeCOM0!(pos0, startatom=1, endatom=size(pos0, 2))

Translate the block pos0[:, startatom:endatom] of coordinates in-place (overwriting it) to have a center of mass of 0, where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$.

See also: makeCOM0

makeCOM0 — Function

makeCOM0(pos0, startatom=1, endatom=size(pos0, 2))

Translate the block pos0[:, startatom:endatom] of coordinates not in-place, returning the result, to have a center of mass of 0, where each column pos0[:, p] has 3 elements representing the Cartesian coordinates of a particle labeled $p$.

See also: makeCOM0!

Data structures and associated functions

TransData — Type

TransData{FT<:AbstractFloat}

Store information relevant to the particular geometrical configuration, not intrinsic to molecules themselves, involving rigid geometrical transformations (translations and rotations); this depends on the number of molecules nummols from MolSystem{T<:OneMol{<:AbstractFloat}}.

numTrans is the number of transformations $t$ (an integer).

transArray is a 3*nummols-by-numTrans real array where each 3-element vector transArray[3*nn-2:3*nn, tt] is the center-of-mass translation vector (relative to the baseline) for molecule nn in transformation labeled tt.

rotArray is a 3*nummols-by-3*numTrans real array where each 3-by-3 block rotArray[3*nn-2:3*nn, 3*tt-2:3*tt] is the center-of-mass rotation matrix (relative to the baseline) for molecule nn in transformation labeled tt.

transLabels is a numTrans-element vector of strings labeling each transformation tt.

changedFromBefore is an nummols-by-numTrans BitArray such that changedFromBefore[nn, tt] is true if molecule nn is transformed in transformation tt relative to transformation tt - 1, false otherwise.

allAtomPos is the horizontal concatenation of atomic coordinates (initially unmodified, later modified, by transformations) across all molecules.

baselineTrans is a 3*nummols-element real vector where each 3-element vector baselineTrans[3*nn-2:3*nn] is the baseline center of mass coordinate for molecule nn.

baselineRot is a 3*nummols-by-3 real matrix where each 3-by-3 block baselineRot[3*nn-2:3*nn,:] is the baseline center of mass rotation matrix for the orientation of molecule nn. Note that for both baselineRot and rotArray, each 3-by-3 Cartesian rotation matrix is given as the transpose of the overall matrix $Z_{1} Y_{2} X_{3}$ in the Tait–Bryan convention, where angle 1 is $\psi$, angle 2 is $\theta$, and angle 3 is $\varphi$. In particular, each rotation matrix is the product $R_{z}(-\psi) R_{y}(-\theta) R_{x}(-\varphi)$, where $R_{i}(\alpha)$ is a standard alibi rotation about Cartesian axis $\vec{e}_{i}$; for example, $R_{z}(\pi/2)$ multiplied on the left of the 3-element column $(1, 0, 0)^{\top}$ yields $(0, 1, 0)^{\top}$, and on the left of $(0, 1, 0)^{\top}$ yields $(-1, 0, 0)^{\top}$. Note also that in the file with name translistfilename, the angles for each baseline orientation as well as each transformation beyond that must be listed in the order phi then theta then psi, in DEGREES.

readTransData — Function

readTransData(myMolData::MolSystem{<:OneMol{FT}},
              translistfilename::AbstractString="") where FT<:AbstractFloat

Construct TransData from existing molecular data myMolData and a file of name translistfilename containing information about geometric transformations; if no file is specified, produce a default set of transformations (no displacements or rotations at all).

See also: MolSystem{<:OneMol{FT}}

initializeAtomPos! — Function

initializeAtomPos!(myTransData::TransData{FT},
                   myMolData::MolSystem{<:OneMol{FT}}) where FT<:AbstractFloat

Initialize atomic coordinates across all molecules by first shifting coordinates of all molecules to have centers of mass of 0, then rotating all molecular coordinates according to baseline orientations of each, then translating to desired baseline centers of mass, modifying myTransData in-place.

See also: transformAtomPos!, untransformAtomPos!

transformAtomPos! — Function

transformAtomPos!(myTransData::TransData{FT},
                  myMolData::MolSystem{<:OneMol{FT}},
                  tt::Integer) where FT<:AbstractFloat

Transform atomic coordinates across all molecules by first shifting coordinates of all molecules to have centers of mass of 0, then rotating all molecular coordinates according to orientations corresponding to transformation tt relative to the baseline, then translating to desired centers of mass corresponding to transformation tt relative to the baseline, modifying myTransData in-place.

See also: initializeAtomPos!, untransformAtomPos!

untransformAtomPos! — Function

untransformAtomPos!(myTransData::TransData{FT},
                    myMolData::MolSystem{<:OneMol{FT}},
                    tt::Integer) where FT<:AbstractFloat

Undo transformation of atomic coordinates across all molecules by first shifting coordinates of all molecules to have centers of mass of 0, then rotating all molecular coordinates according to orientations corresponding to the opposites of transformation tt to return to the baseline, then translating back to the desired baseline centers of mass, modifying myTransData in-place.

See also: initializeAtomPos!, transformAtomPos!